refactor: move enums and constants to separate files  bb-321
