+++ apps/mobile/src/libs/components/deadline-countdown/libs/constants/constants.ts
@@ -1,0 +1,5 @@
+const DEADLINE_OVER = 0;
+
+const TIME_PAD_FILL = "0";
+
+export { DEADLINE_OVER, TIME_PAD_FILL };
