+++ apps/mobile/src/libs/components/deadline-countdown/libs/enums/milliseconds-per-unit.enum.ts
@@ -1,0 +1,7 @@
+const MillisecondsPerUnit = {
+	DAY: 86_400_000,
+	HOUR: 3_600_000,
+	MINUTE: 60_000,
+} as const;
+
+export { MillisecondsPerUnit };
