+++ apps/mobile/src/libs/components/deadline-countdown/libs/enums/time-pad.enum.ts
@@ -1,0 +1,6 @@
+const TimePad = {
+	HOURS: 2,
+	MINUTES: 2,
+} as const;
+
+export { TimePad };
